Move application to a separate module
This commit refactors the code so that the App struct now has a separate
module.
